Bugfix (1451851): When delivering to a building, check whether it is still the same building.
When upgrading, a building changes in place and delivering an item to the
new building - even though the item was never requested by that new building -
will leave the item in an undefined state which eventually causes an exception
because the game cannot transport the item out of the building.

The fix is to have carriers check whether the building is still the same,
just before handing off the item.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2381 37b2a8de-5219-0410-9f54-a31bc463ab9c