Fix bug 1911569 for real - the log message of r3024 contained the wrong bug number.
The issue was how production sites with multiple workers were handled.
In the old code, the main worker was put idle if not all workers were
available. The Bob act change of r3012 caused this to fail.

Putting the main worker idle is problematic because the main worker also
needs to perform maintenance tasks like fetch from flag. So now the main
worker gets the buildingwork task immediately, but is stopped from doing
any kind of "real" work in get_buildingwork().


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3028 37b2a8de-5219-0410-9f54-a31bc463ab9c