Bandaid for Bug 1917914: Rename libsound.a to libwlsound.a
This was suggest by Klaus Michael Mueller on the mailing list, and it
hopefully fixes that bug, at least temporarily.

Of course, the real problem is that *we should not search for those
libraries in system directories*! So this is really just a sloppy hack
around the real problem. Unfortunately, I don't know enough about Scons
to work the necessary magic for a proper fix.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3076 37b2a8de-5219-0410-9f54-a31bc463ab9c