Fix economy rebalance storms.
Simplify the logic that throttles rebalance commands; it's very similar
to what bobs do now. This can temporarily cause more rebalance commands
in the command queue than the old logic, but it's written to be more robust
and should have less potential for failure.

Also introduce Write/Read functions for economies, which should fix problems
with interaction of commands with save/load logic.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3088 37b2a8de-5219-0410-9f54-a31bc463ab9c