Terrain renderer: Fix rounding error in polygon clipping.
Calculating the ty/dty and b/db values is now delayed until the creation
of the edge lists. This fixes a very subtle rounding issue with triangles
that are clipped by the left border of the screen. The error would be
practically invisible during normal play, but it's not so hard to notice
when using the Watch Window to follow a bob.


git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3163 37b2a8de-5219-0410-9f54-a31bc463ab9c