Fix bug #1945549 reported and hunted down by Jari Hautio - jaari. In Critter_Bob::roam_update, the state pointer was used when it might be invalid after a successful call to "bool Bob::start_task_movepath". Now state->ivar is set to 0 before the call to start_task_movepath and then set to 1 again if the call failed.
Attempt to prevent this kind of error in <whatever>_update(Game *, State *) functions by combinging statements that invalidate the state pointer with the return statement (that usually follows immediately). Now it is more difficult to accidentally insert a statement that uses the state pointer between for example a call to pop_task or start_task_<whatever> and the end of control.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3184 37b2a8de-5219-0410-9f54-a31bc463ab9c