Fixed a bug with resource overlays in the editor
When a map was loaded to be edited again, when the number of resources
would been decreased to zero, the resource overlay would not vanish.
The Problem was that upon loading, when the overlays were first
installed, PicMod_Game was used while in the Editor PicMod_Menu was
used. Fixed loading to PicMod_Menu too.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@5026 37b2a8de-5219-0410-9f54-a31bc463ab9c