Disable molog() output by default
This reduces output to the console significantly, which is a good thing if
the console is run by an inefficient terminal emulator.

When opening a debug window for a Map_Object, output of that object's logs
is automatically enabled.

Output can also be reenabled globally by passing the --verbose option when
starting Widelands.