Sanitize richtext in chat messages on the client side
The previous server-side approach was at least partial snake oil, because
a malicious game host could easily remove those checks.

The current checks still do not eliminate richtext entirely from system
messages, because formatting is used in the context of the meta server and
dedicated server, but it should mitigate the scope of potentially accessible
bugs (not that there are any known problematic bugs in the richtext renderer
in the first place, but better safe than sorry).