* bug fix: in openGl code the line width primitiv in draw_line also affect the draw_rect function. The line width is now set to a fixed width of 1 inside of draw_rect.
The other possible solution is to set the line width back to 1
inside the draw_line function.

I choose to set the line width in the draw_rect function, because
this function maybe get a optional width parameter, too. If this happen
you do not need to set the line width back inside of draw_line, as no
other function is affected.