Remove __attribute__((noreturn)) everywhere.
This causes clang to generate incorrect code (and crashes) if the function does actually return.

Even if it is true that a function doesn't return, it could cause crashes in the future if the function is ever modified in such a way that it may return.

Better to get slightly inefficient code (or other fixable warnings) than this horrible-to-debug mess.