Bug 984197 - reread description and comments since last submit and adjusted spacing of buttons such that they are grouped appropriately.
Stop/Go is on the left.

Enhance/Destroy/Dismantle are grouped together with a space either side (left hand space size is fixed to keep position always the same).

Work Area/Debug/Center/Help are on the right.

For Warehouses and other(?) buildings that have no stop/go/enhance/destroy/dismantle buttons, Work Area/Debug/Center are on the left and Help is on the right. It looks a bit odd otherwise.