Check for economy splits lazily
Whenever a road or port is removed, instead of checking for economy splits
immediately we now postpone it until the next request balancing time.

This can save a lot of time when many roads are destroyed at the same time,
especially when leaving a game. As an added benefit, this also speeds up
game loading time, because of the two-phase loading process that is done
for replay support.