Remove navigating with ',' and '.'.
It did not work properly in b18 and 19 - since it did incorrectly distinguish
between a drag and a jump. This also makes me believe that it is a rarely used
feature and instead of doing the big refactoring of moving its logic inside
MapView (which is the only class that can meaningfully reason about what a jump
is) I opted for deleting it and see if players complain.